refactor: split large test files to comply with file length limits
- Split CharacterService.test.ts (795 lines) into focused test files:
  - CharacterService.helpers.ts - shared utilities
  - CharacterService.create.test.ts - create operations (273 lines)
  - CharacterService.query.test.ts - query operations (142 lines)
  - CharacterService.update.test.ts - update operations (191 lines)
  - CharacterService.delete.test.ts - delete operations (67 lines)
  - CharacterService.misc.test.ts - miscellaneous tests (123 lines)

- Split PartyService.test.ts (646 lines) into focused test files:
  - PartyService.helpers.ts - shared utilities
  - PartyService.create.test.ts - create operations (191 lines)
  - PartyService.query.test.ts - query operations (103 lines)
  - PartyService.update.test.ts - update operations (151 lines)
  - PartyService.delete.test.ts - delete operations (105 lines)
  - PartyService.misc.test.ts - miscellaneous tests (84 lines)

- All new test files are under the 450-line Codacy limit
- All 109 tests pass successfully
- Maintained comprehensive test coverage and functionality
