+++ packages/server/src/services/__tests__/CharacterService.misc.test.ts
@@ -1,0 +1,201 @@
+import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
+import { CharacterService, CreateCharacterData, UpdateCharacterData } from '../CharacterService';
+import { 
+  createMockParty, 
+  createMockCharacter, 
+  createExpectedDefaultData,
+  createMockPrisma 
+} from './CharacterService.helpers';
+
+describe('CharacterService - miscellaneous tests', () => {
+  let characterService: CharacterService;
+  let mockPrisma: ReturnType<typeof createMockPrisma>;
+
+  beforeEach(() => {
+    mockPrisma = createMockPrisma();
+    characterService = new CharacterService(mockPrisma);
+    vi.clearAllMocks();
+  });
+
+  afterEach(() => {
+    vi.resetAllMocks();
+  });
+
+  describe('private methods', () => {
+    it('should validate update data correctly', () => {
+      const validData = {
+        name: 'Valid Name',
+        race: 'Valid Race',
+        classes: [{ className: 'Fighter', level: 1 }],
+        currentHp: 10,
+        tempHp: 5
+      };
+
+      // This should not throw
+      expect(() => (characterService as any).validateUpdateData(validData)).not.toThrow();
+    });
+
+    it('should build update data correctly', () => {
+      const inputData: UpdateCharacterData = {
+        name: '  Test Name  ',
+        playerName: '  Test Player  ',
+        race: '  Test Race  ',
+        notes: '  Test Notes  ',
+        level: 5,
+        ac: 16,
+        currentHp: 30
+      };
+
+      const result = (characterService as any).buildUpdateData(inputData);
+
+      expect(result).toEqual({
+        name: 'Test Name',
+        playerName: 'Test Player',
+        race: 'Test Race',
+        notes: 'Test Notes',
+        level: 5,
+        ac: 16,
+        currentHp: 30
+      });
+    });
+
+    it('should process string fields with null handling', () => {
+      const inputData = {
+        playerName: '',
+        notes: '   '
+      };
+      const updateData: any = {};
+
+      (characterService as any).processStringFields(inputData, updateData);
+
+      expect(updateData).toEqual({
+        playerName: null,
+        notes: null
+      });
+    });
+
+    it('should process direct fields correctly', () => {
+      const inputData = {
+        level: 10,
+        ac: 18,
+        abilities: { str: 16, dex: 14, con: 15, int: 12, wis: 13, cha: 11 },
+        features: ['Feature 1', 'Feature 2']
+      };
+      const updateData: any = {};
+
+      (characterService as any).processDirectFields(inputData, updateData);
+
+      expect(updateData).toEqual({
+        level: 10,
+        ac: 18,
+        abilities: { str: 16, dex: 14, con: 15, int: 12, wis: 13, cha: 11 },
+        features: ['Feature 1', 'Feature 2']
+      });
+    });
+  });
+
+  describe('edge cases', () => {
+    it('should handle undefined values in update data', async () => {
+      const updateData: UpdateCharacterData = {
+        currentHp: 0
+      };
+
+      const existingCharacter = createMockCharacter();
+      vi.spyOn(characterService, 'findById').mockResolvedValue(existingCharacter);
+      mockPrisma.character.update = vi.fn().mockResolvedValue(createMockCharacter());
+
+      await characterService.update('char_123', 'user_123', updateData);
+
+      expect(mockPrisma.character.update).toHaveBeenCalledWith({
+        where: { id: 'char_123' },
+        data: {
+          currentHp: 0
+        },
+      });
+    });
+
+    it('should calculate proficiency bonus correctly for different levels', async () => {
+      const testCases = [
+        { level: 1, expectedBonus: 2 },
+        { level: 4, expectedBonus: 2 },
+        { level: 5, expectedBonus: 3 },
+        { level: 8, expectedBonus: 3 },
+        { level: 9, expectedBonus: 4 },
+        { level: 12, expectedBonus: 4 },
+        { level: 13, expectedBonus: 5 },
+        { level: 16, expectedBonus: 5 },
+        { level: 17, expectedBonus: 6 },
+        { level: 20, expectedBonus: 6 }
+      ];
+
+      mockPrisma.party.findFirst = vi.fn().mockResolvedValue(createMockParty());
+
+      for (const testCase of testCases) {
+        const createData: CreateCharacterData = {
+          partyId: 'party_123',
+          name: `Level ${testCase.level} Character`,
+          race: 'Human',
+          classes: [{ className: 'Fighter', level: testCase.level }],
+          level: testCase.level
+        };
+
+        mockPrisma.character.create = vi.fn().mockResolvedValue(createMockCharacter());
+
+        await characterService.create('user_123', createData);
+
+        expect(mockPrisma.character.create).toHaveBeenCalledWith(
+          expect.objectContaining({
+            data: expect.objectContaining({
+              proficiencyBonus: testCase.expectedBonus
+            })
+          })
+        );
+      }
+    });
+
+    it('should handle zero values correctly for tempHp but default speed', async () => {
+      const dataWithZeros: CreateCharacterData = {
+        partyId: 'party_123',
+        name: 'Zero Character',
+        race: 'Human',
+        classes: [{ className: 'Fighter', level: 1 }],
+        tempHp: 0,
+        speed: 0 // Service will default this to 30 since 0 is falsy
+      };
+
+      mockPrisma.party.findFirst = vi.fn().mockResolvedValue(createMockParty());
+      mockPrisma.character.create = vi.fn().mockResolvedValue(createMockCharacter());
+
+      await characterService.create('user_123', dataWithZeros);
+
+      expect(mockPrisma.character.create).toHaveBeenCalledWith({
+        data: {
+          partyId: 'party_123',
+          name: 'Zero Character',
+          playerName: null,
+          race: 'Human',
+          classes: [{ className: 'Fighter', level: 1 }],
+          level: 1,
+          ac: 10,
+          maxHp: 10,
+          currentHp: 10,
+          tempHp: 0, // This should be 0 as provided
+          hitDice: null,
+          speed: 30, // This defaults to 30 because 0 is falsy
+          abilities: {
+            str: 10,
+            dex: 10,
+            con: 10,
+            int: 10,
+            wis: 10,
+            cha: 10
+          },
+          proficiencyBonus: 2,
+          features: [],
+          equipment: [],
+          notes: null,
+        },
+      });
+    });
+  });
+});
